backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
PyMuPDF
langchain-groq
langchain-core
langchain-text-splitters
langgraph
sentence-transformers
numpy
rich
pydantic-settings
requests
httpx
PyJWT[crypto]
